+++ opensis-api/opensisAPI/Controllers/SectionController.cs
@@ -29,7 +29,17 @@ namespace opensisAPI.Controllers
             SectionAddViewModel sectionAdd = new SectionAddViewModel();
             try
             {
-                sectionAdd = _sectionService.SaveSection(section);
+                if (section.tableSections.SchoolId > 0)
+                {
+                    sectionAdd = _sectionService.SaveSection(section);
+                }
+                else
+                {
+                    sectionAdd._token = section._token;
+                    sectionAdd._tenantName = section._tenantName;
+                    sectionAdd._failure = true;
+                    sectionAdd._message = "Please enter valid scholl id";
+                }
             }
             catch (Exception es)
             {
@@ -46,7 +56,17 @@ namespace opensisAPI.Controllers
             SectionAddViewModel sectionUpdate = new SectionAddViewModel();
             try
             {
-                sectionUpdate = _sectionService.UpdateSection(section);
+                if (section.tableSections.SchoolId > 0)
+                {
+                    sectionUpdate = _sectionService.UpdateSection(section);
+                }
+                else
+                {
+                    sectionUpdate._token = section._token;
+                    sectionUpdate._tenantName = section._tenantName;
+                    sectionUpdate._failure = true;
+                    sectionUpdate._message = "Please enter valid scholl id";
+                }
             }
             catch (Exception es)
             {
@@ -64,7 +84,17 @@ namespace opensisAPI.Controllers
             SectionAddViewModel sectionView = new SectionAddViewModel();
             try
             {
-                sectionView = _sectionService.ViewSection(section);
+                if (section.tableSections.SchoolId > 0)
+                {
+                    sectionView = _sectionService.ViewSection(section);
+                }
+                else
+                {
+                    sectionView._token = section._token;
+                    sectionView._tenantName = section._tenantName;
+                    sectionView._failure = true;
+                    sectionView._message = "Please enter valid scholl id";
+                }
             }
             catch (Exception es)
             {
@@ -83,7 +113,17 @@ namespace opensisAPI.Controllers
             SectionListViewModel sectionList = new SectionListViewModel();
             try
             {
-                sectionList = _sectionService.GetAllsection(section);
+                if (section.SchoolId > 0)
+                {
+                    sectionList = _sectionService.GetAllsection(section);
+                }
+                else
+                {
+                    sectionList._token = section._token;
+                    sectionList._tenantName = section._tenantName;
+                    sectionList._failure = true;
+                    sectionList._message = "Please enter valid scholl id";
+                }
             }
             catch (Exception es)
             {
@@ -100,7 +140,17 @@ namespace opensisAPI.Controllers
             SectionAddViewModel sectionDelete = new SectionAddViewModel();
             try
             {
-                sectionDelete = _sectionService.DeleteSection(section);
+                if (section.tableSections.SchoolId > 0)
+                {
+                    sectionDelete = _sectionService.DeleteSection(section);
+                }
+                else
+                {
+                    sectionDelete._token = section._token;
+                    sectionDelete._tenantName = section._tenantName;
+                    sectionDelete._failure = true;
+                    sectionDelete._message = "Please enter valid scholl id";
+                }
             }
             catch (Exception es)
             {
